refactor(bridge): extract forward-message FFI seam
