+++ whatsrust/lib/main.go
@@ -28,12 +28,6 @@ typedef struct {
 } FileMessage;
 
 typedef struct {
-	uint32_t succeeded;
-	uint32_t failed;
-	uint8_t failure;
-} ForwardResult;
-
-typedef struct {
 	uint8_t kind;
 	JID id;
 	char* const* messageIDs;
@@ -586,29 +580,6 @@ func C_TestEmitPresenceEventsConcurrently(from *C.char, count C.uint32_t) {
 		}(index)
 	}
 	wait.Wait()
-}
-
-//export C_ForwardMessage
-func C_ForwardMessage(sourceID *C.char, sourceChat C.JID, sourceSender C.JID, sourceIsFromMe C.bool, destinations **C.char, destinationCount C.size_t, forwardSource *C.uint8_t, forwardSourceLen C.size_t) C.ForwardResult {
-	if sourceID == nil || sourceChat == nil || sourceSender == nil || destinations == nil || destinationCount == 0 {
-		return C.ForwardResult{}
-	}
-	rawDestinations := unsafe.Slice(destinations, int(destinationCount))
-	destinationStrings := make([]string, 0, len(rawDestinations))
-	for _, destination := range rawDestinations {
-		if destination == nil {
-			return C.ForwardResult{failed: C.uint32_t(destinationCount)}
-		}
-		destinationStrings = append(destinationStrings, C.GoString(destination))
-	}
-	return forwardMessages(
-		C.GoString(sourceID),
-		C.GoString(sourceChat),
-		C.GoString(sourceSender),
-		bool(sourceIsFromMe),
-		destinationStrings,
-		forwardingSourceBytes(forwardSource, forwardSourceLen),
-	).cResult()
 }
 
 func main() {} // Required for CGO
